Move save_image to load_neuroimaging_data.py
