+++ FastSurferCNN/data_loader/load_neuroimaging_data.py
@@ -27,6 +27,8 @@ from skimage.measure import label
 from torch.utils.data.dataset import Dataset
 from .conform import is_conform, conform, check_affine_in_nifti
 
+supported_output_file_formats = ['mgz', 'nii', 'nii.gz']
+
 ##
 # Helper Functions
 ##
@@ -67,6 +69,34 @@ def load_and_conform_image(img_filename, interpol=1, logger=None):
     orig = np.asanyarray(orig.dataobj)
 
     return header_info, affine_info, orig
+
+def save_image(pred_prob, affine_info, header_info, save_as):
+    """
+    Save the result of fastsurfercnn, according to the desired output file format.
+    Supported formats are defined in supported_output_file_formats.
+
+    :param numpy.ndarray pred_prob: predictions (in freesurfer label space)
+    :param numpy.ndarray affine_info: image affine information
+    :param nibabel.freesurfer.mghformat.MGHHeader header_info: image header information
+    :param str save_as: name under which to save prediction; this determines output file format
+
+    :return None: saves predictions to save_as
+    """
+
+    assert any(save_as.endswith(file_ext) for file_ext in supported_output_file_formats), \
+            'Output filename does not contain a supported file format (' + ', '.join(file_ext for file_ext in supported_output_file_formats) + ')!'
+
+    mapped_aseg_img = None
+    if save_as.endswith('mgz'):
+        mapped_aseg_img = nib.MGHImage(pred_prob, affine_info, header_info)
+    elif any(save_as.endswith(file_ext) for file_ext in ['nii', 'nii.gz']):
+        mapped_aseg_img = nib.nifti1.Nifti1Pair(pred_prob, affine_info, header_info)
+
+    if any(save_as.endswith(file_ext) for file_ext in ['mgz', 'nii']):
+        nib.save(mapped_aseg_img, save_as)
+    elif save_as.endswith('nii.gz'):
+        ## For correct outputs, nii.gz files should be saved using the nifti1 sub-module's save():
+        nib.nifti1.save(mapped_aseg_img, save_as)
 
 
 # Transformation for mapping
